requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# pyCBT/providers/scrappers/slickcharts.py: 2
# pyCBT/providers/scrappers/wikipedia.py: 2,3
lxml == 4.1.1

# pyCBT/common/timezone.py: 2
# pyCBT/providers/oanda/historical.py: 1
numpy == 1.14.1

# pyCBT/providers/oanda/account.py: 16
# pyCBT/providers/oanda/historical.py: 3,5,6
oandapyV20 == 0.5.0

# pyCBT/providers/eia/series.py: 2
# pyCBT/providers/oanda/historical.py: 2
# pyCBT/providers/scrappers/investing.py: 2
# pyCBT/providers/scrappers/slickcharts.py: 3
# pyCBT/providers/scrappers/wikipedia.py: 4
pandas == 0.22.0

# pyCBT/common/timezone.py: 3
python_dateutil == 2.6.1

# pyCBT/common/timezone.py: 1
pytz == 2018.3

# pyCBT/providers/oanda/account.py: 14
ruamel.yaml == 0.15.35

# pyCBT/providers/scrappers/investing.py: 3,4,5,6,7
selenium == 3.10.0
